Replace Google Protobuf implementation with square-wire implementation: Reduce method count, optimize for android

+++ ZaloPay/domain/src/main/java/vn/com/vng/zalopay/domain/Enums.java
@@ -6,16 +6,16 @@ package vn.com.vng.zalopay.domain;
 public class Enums {
 
     public enum Platform {
-        IOS((short) 1),
-        ANDROID((short) 2),
-        WINDOW_PHONE((short) 3);
-        private short mId;
+        IOS(1),
+        ANDROID(2),
+        WINDOW_PHONE(3);
+        private int mId;
 
-        Platform(short id) {
+        Platform(int id) {
             mId = id;
         }
 
-        public short getId() {
+        public int getId() {
             return mId;
         }
     }
